Resolved Image Upload Issue
The function in the imageRepositiry was using image title to find the image in the data base.Hence created error while there were duplicate title. The was resolved by usung image ID as the Identifier. The corresponding changes in the imageController,imageService, and image.html was done,The issue is now resolved.